+++ app-main/src/main/java/net/causw/app/main/domain/community/report/service/v2/ReportService.java
@@ -5,8 +5,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import net.causw.app.main.domain.community.comment.entity.ChildComment;
 import net.causw.app.main.domain.community.comment.entity.Comment;
-import net.causw.app.main.domain.community.comment.service.v2.implementation.ChildCommentReader;
-import net.causw.app.main.domain.community.comment.service.v2.implementation.CommentReader;
+import net.causw.app.main.domain.community.comment.service.implementation.ChildCommentReader;
+import net.causw.app.main.domain.community.comment.service.implementation.CommentReader;
 import net.causw.app.main.domain.community.post.entity.Post;
 import net.causw.app.main.domain.community.post.service.v2.implementation.PostReader;
 import net.causw.app.main.domain.community.report.entity.Report;
